// Typescript/enums.ts
export enum Elementos {
    Hidrogenio = "hidrogenio",
    Carbono = "carbono",
    Oxigenio = "oxigenio",
    Ferro = "ferro",
    Silicio = "silicio",
}

export enum TipoAreas {
    Planicie = "planicie",
    Montanha = "montanha",
    Floresta = "floresta",
    Deserto = "deserto",
    Caverna = "caverna",
}